examples/demo-1-ai-research-api/ai-agent/requirements.txt
z402-sdk>=0.1.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
rich>=13.0.0
